test: extract skill types suites
