+++ crates/agent-engine/src/skills/types_tests.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[path = "types_availability_tests.rs"]
+mod availability;
+#[path = "types_io_tests.rs"]
+mod io;
+#[path = "types_metadata_tests.rs"]
+mod metadata;
